+++ math.js
@@ -4834,7 +4834,7 @@ diag.doc = {
         'diag(a)'
     ],
     'seealso': [
-        'identity', 'ones', 'range', 'size', 'transpose', 'zeros'
+        'identity', 'ones', 'range', 'size', 'squeeze', 'transpose', 'zeros'
     ]
 };
 /**
@@ -4903,7 +4903,7 @@ identity.doc = {
         'identity(size(a))'
     ],
     'seealso': [
-        'diag', 'ones', 'range', 'size', 'transpose', 'zeros'
+        'diag', 'ones', 'range', 'size', 'squeeze', 'transpose', 'zeros'
     ]
 };
 /**
@@ -4958,7 +4958,7 @@ ones.doc = {
         'ones(size(a))'
     ],
     'seealso': [
-        'diag', 'identity', 'range', 'size', 'transpose', 'zeros'
+        'diag', 'identity', 'range', 'size', 'squeeze', 'transpose', 'zeros'
     ]
 };
 /**
@@ -5015,6 +5015,73 @@ size.doc = {
         'size(1:6)'
     ],
     'seealso': [
+        'diag', 'identity', 'ones', 'range', 'squeeze', 'transpose', 'zeros'
+    ]
+};
+/**
+ * Remove singleton dimensions from a matrix. squeeze(x)
+ * @param {Matrix | Array} x
+ * @return {Matrix | Array} res
+ */
+function squeeze (x) {
+    if (arguments.length != 1) {
+        throw newArgumentsError('squeeze', arguments.length, 1);
+    }
+
+    if (x instanceof Matrix || x instanceof Range) {
+        return _squeezeArray(x.toArray());
+    }
+    else if (x instanceof Array) {
+        return _squeezeArray(clone(x));
+    }
+    else {
+        // scalar
+        return clone(x);
+    }
+}
+
+math.squeeze = squeeze;
+
+/**
+ * Recursively squeeze a multi dimensional array
+ * @param {Array} array
+ * @return {Array} array
+ * @private
+ */
+function _squeezeArray(array) {
+    if (array.length == 1) {
+        // squeeze this array
+        return _squeezeArray(array[0]);
+    }
+    else {
+        // process all childs
+        for (var i = 0, len = array.length; i < len; i++) {
+            var child = array[i];
+            if (child instanceof Array) {
+                array[i] = _squeezeArray(child);
+            }
+        }
+        return array;
+    }
+}
+
+/**
+ * Function documentation
+ */
+squeeze.doc = {
+    'name': 'squeeze',
+    'category': 'Numerics',
+    'syntax': [
+        'squeeze(x)'
+    ],
+    'description': 'Remove singleton dimensions from a matrix.',
+    'examples': [
+        'a = zeros(1,3,2)',
+        'size(squeeze(a))',
+        'b = zeros(3,1,1)',
+        'size(squeeze(b))'
+    ],
+    'seealso': [
         'diag', 'identity', 'ones', 'range', 'transpose', 'zeros'
     ]
 };
@@ -5068,7 +5135,7 @@ zeros.doc = {
         'zeros(size(a))'
     ],
     'seealso': [
-        'diag', 'identity', 'ones', 'range', 'size', 'transpose'
+        'diag', 'identity', 'ones', 'range', 'size', 'squeeze', 'transpose'
     ]
 };
 /**