+++ CHANGELOG.md
@@ -7,7 +7,7 @@ https://github.com/josdejong/mathjs
 *Note: version 0.5 is incompatible with previous versions.*
 
 - Implemented data types Matrix and Range.
-- Implemented methods clone, size, identity.
+- Implemented methods clone, diag, size, squeeze, identity.
 - Changed: moved all type constructors in the namespace under math.type,
   created construction methods for easy object creation. For example, a complex
   value is now created with "math.complex(2, 3)" instead of